backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-core
pydantic
langchain-groq
python-dotenv
twilio
slowapi
bcrypt
python-jose[cryptography]
pytest
supabase
requests
tenacity
langdetect
